refactor: SimpleConfirmPopup index export 수정
