+++ src/components/ConfirmPopup/SimpleConfirmPopup/index.ts
@@ -1,0 +1,1 @@
+export * from './SimpleConfirmPopup.client';
