+++ src/components/ConfirmPopup/index.ts
@@ -1,4 +1,5 @@
 export * from './ConfirmDeleteKeyword';
 export * from './ConfirmUnSave';
 export * from './CopyInvitation';
+export * from './SimpleConfirmPopup';
 export * from './useConfirmPopup';
